feat(dtype): 引入复数精度策略，后端成为 dtype 单一真源
- 新增 aicir/dtypes.py，定义复数精度策略，支持 NumpyBackend、GPUBackend 和 NPUBackend 的默认精度设置。
- 更新各后端以遵循新的 dtype 策略，确保门矩阵构造在最宽精度下进行。
- 新增全局默认精度设置 API：set_default_dtype 和 get_default_dtype。
- 添加 dtype 策略测试，确保后端精度行为符合预期。

+++ aicir/core/gates.py
@@ -29,7 +29,13 @@ from ..ir import (
     instruction_qubits,
 )
 
-_CDTYPE = np.complex64
+#: 门矩阵的**构造**精度，恒为双精度。
+#:
+#: 精度策略见 ``aicir/dtypes.py``：dtype 的单一真源是后端。门矩阵一律在最宽精度下
+#: 构造，再由后端在边界处窄化（``backend.cast`` / ``astype(backend.dtype)``）。
+#: 反过来做是不可逆的——先把 Rx 构造成 complex64、再喂给 complex128 的态，
+#: 精度就永久损失了；而 double→single 的窄化不会损失超出目标精度的信息。
+_CDTYPE = np.complex128
 
 KET_0 = np.array([[1.0 + 0.0j], [0.0 + 0.0j]], dtype=_CDTYPE)
 KET_1 = np.array([[0.0 + 0.0j], [1.0 + 0.0j]], dtype=_CDTYPE)
